+++ EssentailFeed/EssentailFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,25 @@
+//
+//  FeedCachePolicy.swift
+//  EssentailFeed
+//
+//  Created by Vishal Wagh on 05/06/22.
+//
+
+import Foundation
+
+internal final class FeedCachePolicy {
+    private init() {}
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+    
+    private static var maxCacheAgeInDays: Int {
+        return 7
+    }
+    
+    internal static func validate(_ timeStamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timeStamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
